add Dockerfile and Github test
