+++ erddap_compliance/cc_erddap.py
@@ -1,4 +1,3 @@
-from types import TracebackType
 import requests
 from compliance_checker.runner import ComplianceChecker, CheckSuite
 from erddapy import ERDDAP
@@ -6,14 +5,13 @@ from dateutil import parser
 import pandas as pd
 from datetime import timedelta
 import json
-import argparse
 import traceback
 from pathlib import Path
 import os
 import urllib
 from urllib.parse import urlparse
 
-def main(prog_args):
+def cc_erddap(prog_args):
     # print(prog_args)
     erddap_hostname = urlparse(prog_args.erddap_server).netloc
 
@@ -94,17 +92,12 @@ def run_checker(dataset, prog_args, epy):
     if not download_path:
         print("Error in dataset ",dataset["datasetID"])
         return None
-    checker_names = prog_args.standards
-    verbose = prog_args.verbose
-    criteria = "normal"
-    output_format = prog_args.format
 
     # If text format is selected make file extension "txt" instead
     file_ext = "txt" if (prog_args.format == "text") else prog_args.format
-
-    output_filename = os.path.join(prog_args.output_dir,dataset["datasetID"]+'.'+file_ext)
-
-
+    
+    output_filename= os.path.join(prog_args.output_dir,dataset["datasetID"]+'.'+file_ext)
+    
     """
     Inputs to ComplianceChecker.run_checker
 
@@ -113,18 +106,18 @@ def run_checker(dataset, prog_args, epy):
     verbose         Verbosity of the output (0, 1, 2)
     criteria        Determines failure (lenient, normal, strict)
     output_filename Path to the file for output
-    output_format   Format of the output
+    format   Format of the output
 
     @returns                If the tests failed (based on the criteria)
     """
     
     return_value, errors = ComplianceChecker.run_checker(
         download_path,
-        checker_names,
-        verbose,
-        criteria,
-        output_filename=output_filename,
-        output_format=output_format,
+        checker_names= prog_args.standards,
+        verbose = prog_args.verbose,
+        criteria = "normal",
+        output_filename= os.path.join(prog_args.output_dir,dataset["datasetID"]+'.'+file_ext),
+        output_format=prog_args.format,
     )
 
     if return_value:
@@ -134,7 +127,7 @@ def run_checker(dataset, prog_args, epy):
         print("Errors: ", errors)
 
     # Open the JSON output and get the compliance scores
-    if output_format == "json":
+    if prog_args.format == "json":
         with open(output_filename, "r") as fp:
             cc_data = json.load(fp)
             for standard in prog_args.standards:
@@ -166,94 +159,3 @@ def fetch_dataset_sample(prog_args, dataset_id, download_url):
         return None
 
 
-def prep_args(prog_args):
-    """
-    Prepares submitted arguments for use by the rest of the script.
-    """
-    prog_args.standards = prog_args.standards.split(",")
-    # prog_args.time_offset = int(prog_args.time_offset)
-
-    return prog_args
-
-if __name__ == "__main__":
-    raw_args = argparse.ArgumentParser(formatter_class=argparse.ArgumentDefaultsHelpFormatter)
-    raw_args.add_argument(
-        "erddap_server",
-        help="The URL of an ERDDAP instance e.g. https://www.example.com/erddap/",
-    )
-
-    raw_args.add_argument(
-        "-s",
-        "--standards",
-        help='What Compliance Checker standards each dataset should be checked against.  Multiple standards may be specified as a CSV string. A full list of acceptable values may be gathered by running the command "compliance-checker --list-tests"',
-        default="cf:1.6",
-    )
-
-    raw_args.add_argument(
-        "-e",
-        "--exclude",
-        help="List of datasets to exclude from compliance checker",
-        default="",
-    )
-
-    raw_args.add_argument(
-        "--exclude_regex",
-        help="Use regular expressions to filter list of datasets to exclude",
-        action="store_true",
-    )
-
-    raw_args.add_argument(
-        "-f",
-        "--format",
-        help="Set the output format [text,html,json,json_new]",
-        default="html",
-    )
-
-    raw_args.add_argument(
-        "-o",
-        "--output_dir",
-        help=f"Where reports should be written to",
-        default="results",
-    )
-
-    raw_args.add_argument(
-        "-t",
-        "--time_offset",
-        help="A python Timedelta string that specifies the time range of data to retreive from each dataset.  This is to reduce the size of netCDF files being queried from the ERDDAP server, since metadata compliance is what's being audited and not the actual data, we only need 1 or more records to get a valid netCDF file",
-        default="1hour",
-    )
-
-    raw_args.add_argument(
-        "--timeout",
-        help="Number of seconds to wait for a response from the ERDDAP server when downloading a sample of a dataset locally",
-        type=int,
-        default=30,
-    )
-
-    raw_args.add_argument(
-        "-v",
-        "--verbose",
-        help="Passes the desired verbosity flag to the compliance checker library. Acceptable Values: 0, 1, 2.  The higher the value, the more verbose the output.  Default: 0",
-        action="store_true",
-        default=False,
-    )
-
-
-    raw_args.add_argument(
-        "--work",
-        help="Specify the temporary working directory for downloaded sample files.",
-        default="/tmp/cc_erddap",
-    )
-
-    raw_args.add_argument(
-        "--dataset_id",
-        help="Check single dataset",
-    )
-
-    prog_args = raw_args.parse_args()
-
-    # do some pre-processing on the arguments to make them ready for the
-    # script to interpret
-    prog_args = prep_args(prog_args)
-
-    main(prog_args)
