+++ requirements.txt
@@ -1,29 +1,31 @@
 antlr4-python3-runtime==4.7.2
-certifi==2021.10.8
-cf-units==3.0.1.post0
-cftime==1.5.2
-charset-normalizer==2.0.10
+certifi==2022.6.15
+cf-units==3.1.0
+cftime==1.6.1
+charset-normalizer==2.0.12
 compliance-checker==5.0.2
 decorator==5.1.1
 erddapy==1.2.1
 idna==3.3
 isodate==0.6.1
-Jinja2==3.0.3
-lxml==4.7.1
-MarkupSafe==2.0.1
-netCDF4==1.5.8
-numpy==1.22.1
-OWSLib==0.25.0
+Jinja2==3.1.2
+lxml==4.9.1
+MarkupSafe==2.1.1
+netCDF4==1.6.0
+numpy==1.23.1
+OWSLib
 pandas==1.4.0
-pendulum==2.1.2
+pendulum
 pygeoif==0.7
-pyproj==3.3.0
+pyproj
 python-dateutil==2.8.2
-pytz==2021.3
+pytz==2022.1
 pytzdata==2020.1
 PyYAML==6.0
-regex==2022.1.18
+regex==2022.7.9
 requests==2.27.1
+setuptools==58.1.0
+Shapely==1.8.2
 six==1.16.0
-urllib3==1.26.8
-validators==0.18.2
+urllib3==1.26.10
+validators==0.20.0
